requirements.txt
numpy>=1.23
matplotlib>=3.5
pyyaml>=6.0
plotly>=5.0


#pip install -r requirements.txt
